POI_NN_Deployment/requirements.txt
psycopg2~=2.9.6
pip~=21.3.1
wheel~=0.37.1
Pillow~=9.5.0
Jinja2~=3.1.2
setuptools~=60.2.0
packaging~=23.1
zipp~=3.15.0
tensorflow~=2.12.0
numpy~=1.23.5
keras~=2.12.0
Flask~=2.2.3
python-dotenv